refactor: simplify lead screens
