+++ packages/apps/crm/src/components/organisms/Lead/LeadBottom/LeadBottom.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import {EditButton} from '../../../organisms';
+
+const LeadBottom = ({idLead, navigation}) => {
+  return (
+    <EditButton
+      onPress={() =>
+        navigation.navigate('LeadFormScreen', {
+          idLead: idLead,
+        })
+      }
+    />
+  );
+};
+
+export default LeadBottom;
