+++ packages/apps/crm/src/components/organisms/Lead/index.js
@@ -1,0 +1,3 @@
+export {default as LeadBottom} from './LeadBottom/LeadBottom';
+export {default as LeadDropdownCards} from './LeadDropdownCards/LeadDropdownCards';
+export {default as LeadHeader} from './LeadHeader/LeadHeader';
